frontend/requirements.txt
# Streamlit Frontend Requirements

# Core
streamlit==1.39.0

# HTTP requests
requests==2.32.3
httpx==0.27.2

# WebSocket client
websocket-client==1.8.0

# Markdown and HTML
markdown==3.7
beautifulsoup4==4.12.3

# PDF export
fpdf2==2.8.1

# Date/time handling
python-dateutil==2.9.0

